+++ pykern/api/auth_api.py
@@ -1,0 +1,39 @@
+"""?
+
+:copyright: Copyright (c) 2025 RadiaSoft LLC.  All Rights Reserved.
+:license: http://www.apache.org/licenses/LICENSE-2.0.html
+"""
+
+from pykern.pkcollections import PKDict
+from pykern.pkdebug import pkdc, pkdlog, pkdp
+import pykern.api.const
+import pykern.quest
+
+class AuthAPI(pykern.quest.API):
+
+    #: Defaults version number but allows override
+    VERSION = pykern.api.const.AUTH_API_VERSION
+
+    #: cache of local clients
+    _clients = PKDict()
+
+    async def api_authenticate_connection(self, api_args):
+        """Process AuthRequest from server
+
+        api_args:
+            token (str): secret value evaluated by `AuthAPI`
+            version (int): protocol version
+        Args:
+            api_args (PKDict): what to validate
+        Returns:
+            Result: validation result
+        """
+        if (v := api_args.get("version")) != self.VERSION:
+            raise APICallError(f"invalid version={v}, expected={self.VERSION}")
+        if (t := self.token()) is not None and t != args.token:
+            # Do not log token
+            raise pykern.api.util.APIForbidden()
+        return PKDict()
+
+    def token(self):
+        retrun
